requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
openai>=1.0.0
google-generativeai>=0.8.0
gunicorn>=21.0.0
